Several cleanups/improvements: * Add the ValueSplitting to the system. * Address the count operation that clean up the code. * Some cleanup of tests. * Some restructuration of the DynamoDb code. * Implement the dynamo_db tests but also put the failing tests as ignored ones.
