+++ linera-core/src/unit_tests/wasm_client_tests.rs
@@ -50,6 +50,7 @@ async fn test_rocks_db_create_application(wasm_runtime: WasmRuntime) -> Result<(
     run_test_create_application(MakeRocksDbStoreClient::with_wasm_runtime(wasm_runtime)).await
 }
 
+#[ignore]
 #[cfg(feature = "aws")]
 #[cfg_attr(feature = "wasmer", test_case(WasmRuntime::Wasmer ; "wasmer"))]
 #[cfg_attr(feature = "wasmtime", test_case(WasmRuntime::Wasmtime ; "wasmtime"))]
@@ -164,6 +165,7 @@ async fn test_rocks_db_run_application_with_dependency(
     .await
 }
 
+#[ignore]
 #[cfg(feature = "aws")]
 #[cfg_attr(feature = "wasmer", test_case(WasmRuntime::Wasmer ; "wasmer"))]
 #[cfg_attr(feature = "wasmtime", test_case(WasmRuntime::Wasmtime ; "wasmtime"))]
