+++ linera-storage/src/unit_tests/dynamo_db.rs
@@ -8,7 +8,7 @@ use std::mem;
 /// Tests if released guards don't use memory.
 #[tokio::test]
 async fn guards_dont_leak() -> Result<(), anyhow::Error> {
-    let store = DynamoDbStoreClient::new_test().await;
+    let store = DynamoDbStoreClient::make_test_client(None).await;
     let chain_id = ChainId::root(1);
     // There should be no active guards when initialized
     assert_eq!(store.client.guards.active_guards(), 0);
