+++ linera-storage/src/unit_tests/rocks_db.rs
@@ -9,7 +9,7 @@ use std::mem;
 /// Tests if released guards don't use memory.
 #[tokio::test]
 async fn guards_dont_leak() -> Result<(), anyhow::Error> {
-    let store = RocksDbStoreClient::new_test().await;
+    let store = RocksDbStoreClient::make_test_client(None).await;
     let chain_id = ChainId::root(1);
     // There should be no active guards when initialized
     assert_eq!(store.client.guards.active_guards(), 0);
