+++ linera-views/Cargo.toml
@@ -22,6 +22,7 @@ db_timings = []
 metrics = ["dep:hex", "dep:metrics"]
 
 [dependencies]
+anyhow = { workspace = true }
 static_assertions = { workspace = true }
 async-lock = { workspace = true }
 async-trait = { workspace = true }
