+++ linera-views/src/dynamo_db.rs
@@ -6,7 +6,7 @@ use crate::{
     common::{ContextFromDb, KeyIterable, KeyValueIterable, KeyValueStoreClient, MIN_VIEW_TAG},
     localstack,
     lru_caching::LruCachingKeyValueClient,
-    value_splitting::DatabaseConsistencyError,
+    value_splitting::{DatabaseConsistencyError, ValueSplittingKeyValueStoreClient},
 };
 use async_lock::{Semaphore, SemaphoreGuard};
 use async_trait::async_trait;
@@ -25,6 +25,16 @@ use std::{collections::HashMap, mem, str::FromStr, sync::Arc};
 use thiserror::Error;
 
 use static_assertions as sa;
+
+#[cfg(any(test, feature = "test"))]
+use {
+    crate::lru_caching::TEST_CACHE_SIZE,
+    anyhow::{Context, Error},
+    aws_sdk_s3::Endpoint,
+    aws_types::SdkConfig,
+    std::env,
+    tokio::sync::{Mutex, MutexGuard},
+};
 
 /// The configuration to connect to DynamoDB.
 pub type Config = aws_sdk_dynamodb::Config;
@@ -514,8 +524,7 @@ impl KeyValueIterable<DynamoDbContextError> for DynamoDbKeyValues {
 pub struct DynamoDbClientInternal {
     client: Client,
     table: TableName,
-    count: Arc<Semaphore>,
-    max_concurrent_queries: Option<usize>,
+    semaphore: Option<Arc<Semaphore>>,
     max_stream_queries: usize,
 }
 
@@ -534,9 +543,9 @@ impl DeletePrefixExpander for DynamoDbClientInternal {
 impl DynamoDbClientInternal {
     /// Obtains the semaphore lock on the database if needed.
     async fn acquire(&self) -> Option<SemaphoreGuard<'_>> {
-        match self.max_concurrent_queries {
+        match &self.semaphore {
             None => None,
-            Some(_max_concurrent_queries) => Some(self.count.acquire().await),
+            Some(count) => Some(count.acquire().await),
         }
     }
 
@@ -548,13 +557,11 @@ impl DynamoDbClientInternal {
         max_stream_queries: usize,
     ) -> Result<(Self, TableStatus), DynamoDbContextError> {
         let client = Client::from_conf(config.into());
-        let n = max_concurrent_queries.unwrap_or(0);
-        let count = Arc::new(Semaphore::new(n));
+        let semaphore = max_concurrent_queries.map(|n| Arc::new(Semaphore::new(n)));
         let db = DynamoDbClientInternal {
             client,
             table,
-            count,
-            max_concurrent_queries,
+            semaphore,
             max_stream_queries,
         };
         let table_status = db.create_table_if_needed().await?;
@@ -753,15 +760,15 @@ impl KeyValueStoreClient for DynamoDbClientInternal {
 /// A shared DB client for DynamoDb implementing LruCaching
 #[derive(Clone)]
 pub struct DynamoDbClient {
-    client: LruCachingKeyValueClient<DynamoDbClientInternal>,
+    client: LruCachingKeyValueClient<ValueSplittingKeyValueStoreClient<DynamoDbClientInternal>>,
 }
 
 #[async_trait]
 impl KeyValueStoreClient for DynamoDbClient {
     const MAX_VALUE_SIZE: usize = MAX_VALUE_SIZE;
     type Error = DynamoDbContextError;
-    type Keys = DynamoDbKeys;
-    type KeyValues = DynamoDbKeyValues;
+    type Keys = Vec<Vec<u8>>;
+    type KeyValues = Vec<(Vec<u8>, Vec<u8>)>;
 
     fn max_stream_queries(&self) -> usize {
         self.client.max_stream_queries()
@@ -817,16 +824,13 @@ impl DynamoDbClient {
             max_stream_queries,
         )
         .await?;
-        Ok((
-            Self {
-                client: LruCachingKeyValueClient::new(client, cache_size),
-            },
-            table_name,
-        ))
+        let client = ValueSplittingKeyValueStoreClient::new(client);
+        let client = DynamoDbClient {
+            client: LruCachingKeyValueClient::new(client, cache_size),
+        };
+        Ok((client, table_name))
     }
-}
 
-impl DynamoDbClient {
     /// Creates a new [`DynamoDbClientInternal`] instance.
     pub async fn new(
         table: TableName,
@@ -1164,4 +1168,162 @@ impl IsResourceInUseException for SdkError<aws_sdk_dynamodb::error::CreateTableE
             }
         )
     }
+}
+
+/// A static lock to prevent multiple tests from using the same LocalStack instance at the same
+/// time.
+#[cfg(any(test, feature = "test"))]
+static LOCALSTACK_GUARD: Mutex<()> = Mutex::const_new(());
+
+/// Name of the environment variable with the address to a LocalStack instance.
+#[cfg(any(test, feature = "test"))]
+const LOCALSTACK_ENDPOINT: &str = "LOCALSTACK_ENDPOINT";
+
+/// A type to help tests that need a LocalStack instance.
+#[cfg(any(test, feature = "test"))]
+pub struct LocalStackTestContext {
+    base_config: SdkConfig,
+    endpoint: Endpoint,
+    _guard: MutexGuard<'static, ()>,
+}
+
+#[cfg(any(test, feature = "test"))]
+impl LocalStackTestContext {
+    /// Creates an instance of [`LocalStackTestContext`], loading the necessary LocalStack
+    /// configuration.
+    ///
+    /// An address to the LocalStack instance must be specified using a `LOCALSTACK_ENDPOINT`
+    /// environment variable.
+    ///
+    /// This also locks the `LOCALSTACK_GUARD` to enforce that only one test has access to the
+    /// LocalStack instance.
+    pub async fn new() -> Result<LocalStackTestContext, Error> {
+        let base_config = aws_config::load_from_env().await;
+        let endpoint = Self::load_endpoint()?;
+        let _guard = LOCALSTACK_GUARD.lock().await;
+
+        let context = LocalStackTestContext {
+            base_config,
+            endpoint,
+            _guard,
+        };
+
+        context.clear().await?;
+
+        Ok(context)
+    }
+
+    /// Creates an [`Endpoint`] using the configuration in the [`LOCALSTACK_ENDPOINT`] environment
+    /// variable.
+    fn load_endpoint() -> Result<Endpoint, Error> {
+        let endpoint_address = env::var(LOCALSTACK_ENDPOINT)
+            .with_context(|| {
+                format!(
+                    "Missing LocalStack endpoint address in {LOCALSTACK_ENDPOINT:?} \
+                    environment variable"
+                )
+            })?
+            .parse()
+            .context("LocalStack endpoint address is not a valid URI")?;
+
+        Ok(Endpoint::immutable(endpoint_address))
+    }
+
+    /// Creates a new [`aws_sdk_s3::Config`] for tests, using a LocalStack instance.
+    pub fn s3_config(&self) -> aws_sdk_s3::Config {
+        aws_sdk_s3::config::Builder::from(&self.base_config)
+            .endpoint_resolver(self.endpoint.clone())
+            .build()
+    }
+
+    /// Creates a new [`aws_sdk_dynamodb::Config`] for tests, using a LocalStack instance.
+    pub fn dynamo_db_config(&self) -> aws_sdk_dynamodb::Config {
+        aws_sdk_dynamodb::config::Builder::from(&self.base_config)
+            .endpoint_resolver(self.endpoint.clone())
+            .build()
+    }
+
+    /// Removes all stored data from LocalStack storage.
+    async fn clear(&self) -> Result<(), Error> {
+        self.remove_buckets().await?;
+        self.remove_tables().await?;
+
+        Ok(())
+    }
+
+    /// Removes all buckets from the LocalStack S3 storage.
+    async fn remove_buckets(&self) -> Result<(), Error> {
+        let client = aws_sdk_s3::Client::from_conf(self.s3_config());
+
+        for bucket in list_buckets(&client).await.unwrap_or_default() {
+            let objects = client.list_objects().bucket(&bucket).send().await?;
+
+            for object in objects.contents().into_iter().flatten() {
+                if let Some(key) = object.key.as_ref() {
+                    client
+                        .delete_object()
+                        .bucket(&bucket)
+                        .key(key)
+                        .send()
+                        .await?;
+                }
+            }
+
+            client.delete_bucket().bucket(bucket).send().await?;
+        }
+
+        Ok(())
+    }
+
+    /// Removes all tables from the LocalStack DynamoDB storage.
+    async fn remove_tables(&self) -> Result<(), Error> {
+        let client = aws_sdk_dynamodb::Client::from_conf(self.dynamo_db_config());
+
+        for table in list_tables(&client).await.unwrap_or_default() {
+            client.delete_table().table_name(table).send().await?;
+        }
+
+        Ok(())
+    }
+}
+
+/// Creates a basic client that can be used for tests.
+#[cfg(any(test, feature = "test"))]
+pub async fn create_dynamo_db_test_client() -> DynamoDbClient {
+    let localstack = LocalStackTestContext::new().await.expect("localstack");
+    let (key_value_operation, _) = DynamoDbClient::from_config(
+        localstack.dynamo_db_config(),
+        "test_table".parse().expect("Invalid table name"),
+        Some(TEST_DYNAMO_DB_MAX_CONCURRENT_QUERIES),
+        TEST_DYNAMO_DB_MAX_STREAM_QUERIES,
+        TEST_CACHE_SIZE,
+    )
+    .await
+    .expect("key_value_operation");
+    key_value_operation
+}
+
+/// Helper function to list the names of buckets registered on S3.
+#[cfg(any(test, feature = "test"))]
+async fn list_buckets(client: &aws_sdk_s3::Client) -> Result<Vec<String>, Error> {
+    Ok(client
+        .list_buckets()
+        .send()
+        .await?
+        .buckets
+        .expect("List of buckets was not returned")
+        .into_iter()
+        .filter_map(|bucket| bucket.name)
+        .collect())
+}
+
+/// Helper function to list the names of tables registered on DynamoDB.
+#[cfg(any(test, feature = "test"))]
+async fn list_tables(client: &aws_sdk_dynamodb::Client) -> Result<Vec<String>, Error> {
+    Ok(client
+        .list_tables()
+        .send()
+        .await?
+        .table_names
+        .expect("List of tables was not returned"))
 }
