+++ linera-views/src/test_utils/mod.rs
@@ -9,176 +9,6 @@ use crate::batch::{
 use rand::{Rng, RngCore};
 use std::collections::HashSet;
 
-#[cfg(feature = "aws")]
-use {
-    crate::dynamo_db::DynamoDbClient,
-    crate::dynamo_db::{TEST_DYNAMO_DB_MAX_CONCURRENT_QUERIES, TEST_DYNAMO_DB_MAX_STREAM_QUERIES},
-    crate::lru_caching::TEST_CACHE_SIZE,
-    anyhow::{Context, Error},
-    aws_sdk_s3::Endpoint,
-    aws_types::SdkConfig,
-    std::env,
-    tokio::sync::{Mutex, MutexGuard},
-};
-
-#[cfg(feature = "aws")]
-/// A static lock to prevent multiple tests from using the same LocalStack instance at the same
-/// time.
-static LOCALSTACK_GUARD: Mutex<()> = Mutex::const_new(());
-
-#[cfg(feature = "aws")]
-/// Name of the environment variable with the address to a LocalStack instance.
-const LOCALSTACK_ENDPOINT: &str = "LOCALSTACK_ENDPOINT";
-
-#[cfg(feature = "aws")]
-/// A type to help tests that need a LocalStack instance.
-pub struct LocalStackTestContext {
-    base_config: SdkConfig,
-    endpoint: Endpoint,
-    _guard: MutexGuard<'static, ()>,
-}
-
-#[cfg(feature = "aws")]
-impl LocalStackTestContext {
-    /// Creates an instance of [`LocalStackTestContext`], loading the necessary LocalStack
-    /// configuration.
-    ///
-    /// An address to the LocalStack instance must be specified using a `LOCALSTACK_ENDPOINT`
-    /// environment variable.
-    ///
-    /// This also locks the `LOCALSTACK_GUARD` to enforce that only one test has access to the
-    /// LocalStack instance.
-    pub async fn new() -> Result<LocalStackTestContext, Error> {
-        let base_config = aws_config::load_from_env().await;
-        let endpoint = Self::load_endpoint()?;
-        let _guard = LOCALSTACK_GUARD.lock().await;
-
-        let context = LocalStackTestContext {
-            base_config,
-            endpoint,
-            _guard,
-        };
-
-        context.clear().await?;
-
-        Ok(context)
-    }
-
-    /// Creates an [`Endpoint`] using the configuration in the [`LOCALSTACK_ENDPOINT`] environment
-    /// variable.
-    fn load_endpoint() -> Result<Endpoint, Error> {
-        let endpoint_address = env::var(LOCALSTACK_ENDPOINT)
-            .with_context(|| {
-                format!(
-                    "Missing LocalStack endpoint address in {LOCALSTACK_ENDPOINT:?} \
-                    environment variable"
-                )
-            })?
-            .parse()
-            .context("LocalStack endpoint address is not a valid URI")?;
-
-        Ok(Endpoint::immutable(endpoint_address))
-    }
-
-    /// Creates a new [`aws_sdk_s3::Config`] for tests, using a LocalStack instance.
-    pub fn s3_config(&self) -> aws_sdk_s3::Config {
-        aws_sdk_s3::config::Builder::from(&self.base_config)
-            .endpoint_resolver(self.endpoint.clone())
-            .build()
-    }
-
-    /// Creates a new [`aws_sdk_dynamodb::Config`] for tests, using a LocalStack instance.
-    pub fn dynamo_db_config(&self) -> aws_sdk_dynamodb::Config {
-        aws_sdk_dynamodb::config::Builder::from(&self.base_config)
-            .endpoint_resolver(self.endpoint.clone())
-            .build()
-    }
-
-    /// Removes all stored data from LocalStack storage.
-    async fn clear(&self) -> Result<(), Error> {
-        self.remove_buckets().await?;
-        self.remove_tables().await?;
-
-        Ok(())
-    }
-
-    /// Removes all buckets from the LocalStack S3 storage.
-    async fn remove_buckets(&self) -> Result<(), Error> {
-        let client = aws_sdk_s3::Client::from_conf(self.s3_config());
-
-        for bucket in list_buckets(&client).await.unwrap_or_default() {
-            let objects = client.list_objects().bucket(&bucket).send().await?;
-
-            for object in objects.contents().into_iter().flatten() {
-                if let Some(key) = object.key.as_ref() {
-                    client
-                        .delete_object()
-                        .bucket(&bucket)
-                        .key(key)
-                        .send()
-                        .await?;
-                }
-            }
-
-            client.delete_bucket().bucket(bucket).send().await?;
-        }
-
-        Ok(())
-    }
-
-    /// Removes all tables from the LocalStack DynamoDB storage.
-    async fn remove_tables(&self) -> Result<(), Error> {
-        let client = aws_sdk_dynamodb::Client::from_conf(self.dynamo_db_config());
-
-        for table in list_tables(&client).await.unwrap_or_default() {
-            client.delete_table().table_name(table).send().await?;
-        }
-
-        Ok(())
-    }
-}
-
-/// Creates a basic client that can be used for tests.
-#[cfg(feature = "aws")]
-pub async fn create_dynamo_db_test_client() -> DynamoDbClient {
-    let localstack = LocalStackTestContext::new().await.expect("localstack");
-    let (key_value_operation, _) = DynamoDbClient::from_config(
-        localstack.dynamo_db_config(),
-        "test_table".parse().expect("Invalid table name"),
-        Some(TEST_DYNAMO_DB_MAX_CONCURRENT_QUERIES),
-        TEST_DYNAMO_DB_MAX_STREAM_QUERIES,
-        TEST_CACHE_SIZE,
-    )
-    .await
-    .expect("key_value_operation");
-    key_value_operation
-}
-
-#[cfg(feature = "aws")]
-/// Helper function to list the names of buckets registered on S3.
-pub async fn list_buckets(client: &aws_sdk_s3::Client) -> Result<Vec<String>, Error> {
-    Ok(client
-        .list_buckets()
-        .send()
-        .await?
-        .buckets
-        .expect("List of buckets was not returned")
-        .into_iter()
-        .filter_map(|bucket| bucket.name)
-        .collect())
-}
-
-#[cfg(feature = "aws")]
-/// Helper function to list the names of tables registered on DynamoDB.
-pub async fn list_tables(client: &aws_sdk_dynamodb::Client) -> Result<Vec<String>, Error> {
-    Ok(client
-        .list_tables()
-        .send()
-        .await?
-        .table_names
-        .expect("List of tables was not returned"))
-}
-
 /// Shuffles the values entries randomly
 pub fn random_shuffle<R: RngCore, T: Clone>(rng: &mut R, values: &mut Vec<T>) {
     let n = values.len();
