+++ linera-views/src/unit_tests/dynamo_db_context_tests.rs
@@ -3,11 +3,30 @@
 
 use super::{DynamoDbContext, TableName, TableStatus};
 use crate::{
-    dynamo_db::{TEST_DYNAMO_DB_MAX_CONCURRENT_QUERIES, TEST_DYNAMO_DB_MAX_STREAM_QUERIES},
+    dynamo_db::{
+        list_tables, DynamoDbContextError, LocalStackTestContext,
+        TEST_DYNAMO_DB_MAX_CONCURRENT_QUERIES, TEST_DYNAMO_DB_MAX_STREAM_QUERIES,
+    },
     lru_caching::TEST_CACHE_SIZE,
-    test_utils::{list_tables, LocalStackTestContext},
 };
 use anyhow::Error;
+
+async fn get_table_status(
+    localstack: &LocalStackTestContext,
+    table: &TableName,
+) -> Result<TableStatus, DynamoDbContextError> {
+    let (_storage, table_status) = DynamoDbContext::from_config(
+        localstack.dynamo_db_config(),
+        table.clone(),
+        Some(TEST_DYNAMO_DB_MAX_CONCURRENT_QUERIES),
+        TEST_DYNAMO_DB_MAX_STREAM_QUERIES,
+        TEST_CACHE_SIZE,
+        vec![],
+        (),
+    )
+    .await?;
+    Ok(table_status)
+}
 
 /// Tests if the table for the storage is created when needed.
 #[tokio::test]
@@ -19,16 +38,7 @@ async fn table_is_created() -> Result<(), Error> {
     let initial_tables = list_tables(&client).await?;
     assert!(!initial_tables.contains(table.as_ref()));
 
-    let (_storage, table_status) = DynamoDbContext::from_config(
-        localstack.dynamo_db_config(),
-        table.clone(),
-        Some(TEST_DYNAMO_DB_MAX_CONCURRENT_QUERIES),
-        TEST_DYNAMO_DB_MAX_STREAM_QUERIES,
-        TEST_CACHE_SIZE,
-        vec![],
-        (),
-    )
-    .await?;
+    let table_status = get_table_status(&localstack, &table).await?;
 
     let tables = list_tables(&client).await?;
     assert!(tables.contains(table.as_ref()));
@@ -49,26 +59,8 @@ async fn separate_tables_are_created() -> Result<(), Error> {
     assert!(!initial_tables.contains(first_table.as_ref()));
     assert!(!initial_tables.contains(second_table.as_ref()));
 
-    let (_storage, first_table_status) = DynamoDbContext::from_config(
-        localstack.dynamo_db_config(),
-        first_table.clone(),
-        Some(TEST_DYNAMO_DB_MAX_CONCURRENT_QUERIES),
-        TEST_DYNAMO_DB_MAX_STREAM_QUERIES,
-        TEST_CACHE_SIZE,
-        vec![],
-        (),
-    )
-    .await?;
-    let (_storage, second_table_status) = DynamoDbContext::from_config(
-        localstack.dynamo_db_config(),
-        second_table.clone(),
-        Some(TEST_DYNAMO_DB_MAX_CONCURRENT_QUERIES),
-        TEST_DYNAMO_DB_MAX_STREAM_QUERIES,
-        TEST_CACHE_SIZE,
-        vec![],
-        (),
-    )
-    .await?;
+    let first_table_status = get_table_status(&localstack, &first_table).await?;
+    let second_table_status = get_table_status(&localstack, &second_table).await?;
 
     let tables = list_tables(&client).await?;
     assert!(tables.contains(first_table.as_ref()));
